TK-252 Always build a metrics registry, deprecate metrics.enabled
Previously we have allowed the user to disable the creation of a
MetricsRegistry with the `metrics.enabled` option, though this means
the developer had to guard against a nil registry in all codepaths.

Metrics collection is very light and so here we make the change to
always create a MetricsRegisty, and log a warning message if
`metrics.enabled` is provided.