TK-253 become a wrapper for metrics-clojure's default-registry
Here we depend on the metrics-clojure wrapper for Dropwizard Metrics and use
the global MetricsRegistry that it creates rather than construct our own.

This allows a hybrid usage pattern where metrics may be declared globally and
used without threading the tk-metrics MetricsRegistry though the application,
but we are still able to configure JMX Publishing via service configuration files.