(TK-404) (SPIKE) Add jolokia api as metrics/v2
This patch adds the jolokia-core library to the dependency list and
mounts its servlet at 'metrics/v2' when the `metrics-webservice` is in
use. Jolokia is a small library that wraps a REST API around JMX and
provides great features such as globbing, attribute extraction and much
more.

Example query using globbing and parameter extraction:

    curl -k 'https://localhost:8140/metrics/v2/read/java.lang:name=*,type=GarbageCollector/CollectionCount,CollectionTime'

**CAVEATS**

  - The Jolokia API also exposes JMX control functionality, so "write"
    and "exec" operations are currently restricted by a static
    authorization configuration in `resources/jolokia-access.xml`.

  - The servlet is currently mounted directly using the WebserverService
    add-servlet-handler method, instead of going through the comidi
    routes defined in:

        puppetlabs.trapperkeeper.services.metrics.metrics-core