(TK-394) Accept keyword or string metric domains.
The PE implementation of this protocol enforces that only keywords be
used for the metric domains, whereas this implementation takes strings.
In the future both implementations will only take keywords, since the
domain is used as a lookup key so a keyword is a better fit. In light of
this, change this implementation to take keywords or strings so it's
easy for consumers to use this implementation or the PE one, and in the
future we can make the breaking change of only accepting keyword
domains.