(MAINT) Remove explicit ring-core, servlet-api, and slf4j-api deps
This commit removes the following explicit dependencies from the
project.clj file since these are not referenced anywhere in the
tk-metrics code:

* ring/ring-core
* javax.servlet/javax.servlet-api
* org.slf4j/slf4j-api

An exclusion is explicitly used for the javax.servlet-api dependency
from ring/ring-defaults since the ring-defaults version being inherited
is known to bring in a version of javax.servlet-api which we don't want.
The servlet-api dependency will be picked up from clj-parent /
ring-middleware via ring 1.4.0, which pulls in javax.servlet-api 3.1.0.