(TK-427) Fix request stream closed before parsing
Cheshire's parse-stream will return a lazy-seq if the top level JSON
structure is an array. Currently the code calls parse-stream inside a
with-open block and for larger arrays this will cause the lazy-seq to
be realized after the stream has been closed causing an
exception. This commit just wraps that in a doall to ensure the seq is
realized before closing the stream.