(TK-394) Pass in domain as keyword only
Update tk-metrics to only take domains as keywords

The metrics service originally only accepted domains as strings,
and then eventually it was decided keywords would be better, and
a PR was created to make the registry functions take in either
strings or keywords to maintain backward compatability until we
decided to make the breaking change.

That moment has arrived, and so this commit removes the last
references to domains as strings.