(maint) Remove bad test
Previously, the tk-metrics service didn't have the
`initialize-registry-settings` service function implemented, and the service
would throw an exception if you tried to call it. However, it was recently
updated so that the function is implemented, but the test for this behavior
was still present. The test was incorrectly passing because it was now
throwing a schema error. This commit removes the bad test.