(TK-436) Add support for appending registry settings
This commits changes the behavior of initialize-registry-settings to allow it
to be called more than once for one domain, allowing multiple services to add
in the metrics they would like to whitelist for their domain.

As before, the first call will create the settings. Additional calls with the
same domain will merge and append the new settings to the old settings.

Since the only registry setting is currently :default-metrics-allowed, this
allows multiple TK services to call this service function, adding in their own
list of allowed metrics. Before, only one call could be made per domain, which
meant that you would have to designate a service to be the one that knows about
the list of metrics and initialize them.