(TK-488) Disable Jolokia debug mode by default
This commit flips the default for Jolokia's debug mode from "true"
to "false". This change means that Jolokia will no longer produce
error messages for "normal" request failures like a lookup for
a non-existent JMX entry that returns a 404. These errors could
end up flooding the logs with spurious backtraces when metrics
collection systems were polling for events that had not occurred
yet in the process lifetime (and thus had not created the associated JMX
entry).

This change also means that debug-level messages will not be available
unless metrics.conf is updated to flip the debug setting back to "true".