(maint) Make BC a provided dep and add fips profile
A recent commit added Bouncy Castle as a top-level dependency of this
library, along with updating clj-parent by several major versions.
Pulling in BC this way is not valid for FIPS mode, and will cause issues
when this library is used inside another project that expects to pull in
FIPS BC instead. This updates the profiles in this library to correctly
pull in either normal or FIPS BC for testing, depending on the mode, and
list BC as a provided dep, so that we don't require it directly, but
instead allow it to be specified according to the needs of the consuming
project or production installation.