(TK-497) Allow requiring authorization via config setting
This commit addresses PDB-5261 which documents a transient bug that
allows the metrics endpoint to be accessed without authorization when
the trapperkeeper-metrics bootstrap config contains the authorization
service. The root of this problem is likely a race condition within
trapperkeeper.

This patch adds a trapperkeeper-metrics config option
metrics.metrics-webservice.jolokia.require-auth which defaults to true.
In order to use the trapperkeeper-metrics service without authorization,
this option must explicitly be set to false. When this option is set to
true and the authorization service is not loaded, the
IllegalArgumentException thrown by trapper-keeper's get-service method
is purposefully unhandled and crashes the program. Helpful log messages
are printed when this occurs.