(PE-37632) use tk-ws-jetty10 from clj-parent
In order to move to centralized management of trapperkeeper-webserver-jetty-10, this
moves to clj-parent 7.3.6, and removes the version pin from tk-ws-jetty-10

Other outdated dependencies were also updated.

Additionally, profile declarations that mixed keys of other profiles and
maps were refactored to just include keys as the previous format is now
deprecated and not recommended.